refactor: tap interface into own c file
