+++ wasm/Filelists.mk
@@ -1,3 +1,4 @@
 SRC_FILES= \
 	$(SRC_DIR)/stack.c \
+	$(SRC_DIR)/tap_interface.c \
 	$(SRC_DIR)/arch.c
