+++ wasm/macros.h
@@ -1,0 +1,6 @@
+#ifndef MACROS_H
+#define MACROS_H
+
+#define EXPORT(name) __attribute__((export_name(name)))
+
+#endif  // MACROS_H
